# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ME Piston Component Running Hours Record API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vessels/{vesselId}/components:
    get:
      operationId: listComponents
      tags: [components]
      summary: List all piston components for a vessel
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Component list with live RH computed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
    post:
      operationId: createComponent
      tags: [components]
      summary: Register a new component
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
        "400":
          description: Validation error
  /vessels/{vesselId}/components/{componentId}:
    get:
      operationId: getComponent
      tags: [components]
      summary: Get a single component
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
        "404":
          description: Not found
    patch:
      operationId: updateComponent
      tags: [components]
      summary: Update a component
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
    delete:
      operationId: deleteComponent
      tags: [components]
      summary: Delete a component
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Cylinders ─────────────────────────────────────────────────────────────
...
  /vessels/{vesselId}/component-type-thresholds:
    get:
      operationId: listComponentTypeThresholds
      tags: [component-type-thresholds]
      summary: List all component-type thresholds for a vessel
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of component type thresholds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComponentTypeThreshold"
    post:
      operationId: createComponentTypeThreshold
      tags: [component-type-thresholds]
      summary: Create or upsert a component-type threshold
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentTypeThresholdInput"
      responses:
        "201":
          description: Created/upserted threshold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeThreshold"
  /vessels/{vesselId}/component-type-thresholds/{id}:
    patch:
      operationId: updateComponentTypeThreshold
      tags: [component-type-thresholds]
      summary: Update overhaul/warning RH for a component-type threshold
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentTypeThresholdUpdate"
      responses:
        "200":
          description: Updated threshold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentTypeThreshold"
    delete:
      operationId: deleteComponentTypeThreshold
      tags: [component-type-thresholds]
      summary: Delete a component-type threshold
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Auth ───────────────────────────────────────────────────────────────────
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in — sets an httpOnly session cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated

  # ─── Users (Technical Office only) ────────────────────────────────────────────
...
components:
  schemas:
...
    Component:
      type: object
      required: [componentId, vesselId, componentType, condition, currentStatus, currentLocation, totalAccumulatedRh, createdAt]
      properties:
        componentId:
          type: string
        vesselId:
          type: integer
        componentType:
          type: string
        condition:
          type: string
        currentStatus:
          type: string
        currentLocation:
          type: string
        totalAccumulatedRh:
          type: number
        fittedAtMeRh:
          type: ["number", "null"]
        overhaulRh:
          type: ["integer", "null"]
          description: "Per-component overhaul RH limit (overrides vessel default when set)"
        warningRh:
          type: ["integer", "null"]
          description: "Per-component warning RH threshold (overrides vessel default when set)"
        effectiveOverhaulRh:
          type: integer
          description: "Resolved overhaul RH: individual → type-level → vessel default"
        effectiveWarningRh:
          type: integer
          description: "Resolved warning RH: individual → type-level → vessel default"
        remarks:
          type: ["string", "null"]
        liveRh:
          type: number
          description: "Live total RH = stored + (current_me - fitted_at) if in service"
        alertStatus:
          type: string
        createdAt:
          type: string
    ComponentInput:
      type: object
      required: [componentId, componentType, condition, currentStatus]
      properties:
        componentId:
          type: string
          minLength: 1
        componentType:
          type: string
        condition:
          type: string
        currentStatus:
          type: string
        currentLocation:
          type: string
        totalAccumulatedRh:
          type: number
          default: 0
        fittedAtMeRh:
          type: ["number", "null"]
        overhaulRh:
          type: ["integer", "null"]
          description: "Per-component overhaul RH limit (overrides vessel default when set)"
        warningRh:
          type: ["integer", "null"]
          description: "Per-component warning RH threshold (overrides vessel default when set)"
        remarks:
          type: string
    ComponentUpdate:
      type: object
      properties:
        componentType:
          type: string
        condition:
          type: string
        currentStatus:
          type: string
        currentLocation:
          type: string
        totalAccumulatedRh:
          type: number
        fittedAtMeRh:
          type: ["number", "null"]
        overhaulRh:
          type: ["integer", "null"]
        warningRh:
          type: ["integer", "null"]
        remarks:
          type: string

    # ─── Cylinder ────────────────────────────────────────────────────────────
...
    ComponentTypeThreshold:
      type: object
      required: [id, vesselId, category, componentType, overhaulRh, warningRh]
      properties:
        id:
          type: integer
        vesselId:
          type: integer
        category:
          type: string
          description: "Grouping category, e.g. Piston, Fuel Valve, Exhaust Valve"
        componentType:
          type: string
          description: "Specific component type, e.g. Piston Crown, Piston Skirt"
        overhaulRh:
          type: integer
        warningRh:
          type: integer
    ComponentTypeThresholdInput:
      type: object
      required: [category, componentType, overhaulRh, warningRh]
      properties:
        category:
          type: string
        componentType:
          type: string
        overhaulRh:
          type: integer
          minimum: 0
        warningRh:
          type: integer
          minimum: 0
    ComponentTypeThresholdUpdate:
      type: object
      properties:
        overhaulRh:
          type: integer
          minimum: 0
        warningRh:
          type: integer
          minimum: 0

    # ─── Auth & Users ────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserProfile:
      type: object
      required: [id, email, fullName, role, isActive]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [vessel_officer, technical_office]
        isActive:
          type: boolean
        vesselIds:
          type: array
          items:
            type: integer
          description: "Accessible vessel ids; omitted/empty for technical_office (implicit fleet-wide access)"
...